backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
Pillow==10.1.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.1
requests==2.31.0
python-dotenv==1.0.0
setuptools==69.0.3
